Kreado de apo 'core' kun restrukturigo.
