+++ core/templates/core/home.html
@@ -103,7 +103,7 @@
         <a href="skype:pasporta.servo?chat" title="{% trans "Contact us via Skype" %}" class="highlight-custom">
             <img src="{% static 'img/home/social_skype-sm.png' %}" alt="[skype]"/>
         </a>
-        <a href="mailto:pasportaservo [cxe] tejo.org" title="{% trans "Contact us via email" %}" class="highlight-custom">
+        <a href="mailto:saluton [cxe] pasportaservo.org" title="{% trans "Contact us via email" %}" class="highlight-custom">
             <img src="{% static 'img/home/social_email-sm.png' %}" alt="[email]"/>
         </a>
         <a href="https://trello.com/b/kerCH0sk/pasporta-servo-novaj-funkcioj" target="_blank" title="{% trans "Your ideas and suggestions" %}: Trello" class="highlight-custom narrow">
