+++ hosting/forms.py
@@ -7,52 +7,13 @@ try:
 except ImportError:
     from django.utils.functional import keep_lazy_text
     format_lazy = keep_lazy_text(lambda s, *args, **kwargs: s.format(*args, **kwargs))
-from django.contrib.auth.forms import AuthenticationForm as UserLoginForm, UserCreationForm
 from django.contrib.auth import get_user_model
 
-from django_countries import countries
-from phonenumber_field.formfields import PhoneNumberField
-
-from .models import Profile, Place, Phone, Condition
+from .models import Profile, Place, Phone
 from .validators import TooNearPastValidator, client_side_validated
 from .widgets import ClearableWithPreviewImageInput
 
-
 User = get_user_model()
-
-
-class UserRegistrationForm(UserCreationForm):
-    email = forms.EmailField(label=_("Email"), max_length=254)
-    # Honeypot:
-    name = forms.CharField(widget=forms.TextInput(attrs={'autocomplete':'off'}),
-        help_text=_("Leave blank"), required=False)
-
-    def __init__(self, *args, **kwargs):
-        super(UserRegistrationForm, self).__init__(*args, **kwargs)
-        User._meta.get_field('email')._unique = True
-        for fieldname in ['username', 'password1', 'password2', 'email']:
-            self.fields[fieldname].help_text = None
-            self.fields[fieldname].widget.attrs['placeholder'] = self.fields[fieldname].label
-
-    def clean_email(self):
-        email = self.cleaned_data['email']
-        if email and User.objects.filter(email=email):
-            raise forms.ValidationError(_("User address already in use."))
-        return email
-
-    def clean_name(self):
-        """Remove flies from the honeypot."""
-        flies = self.cleaned_data['name']
-        if flies:
-            raise forms.ValidationError("")
-        return flies
-
-    def save(self, commit=True):
-        user = super(UserRegistrationForm, self).save(commit=False)
-        user.email = self.cleaned_data['email']
-        if commit:
-            user.save()
-        return user
 
 
 @client_side_validated
@@ -302,7 +263,7 @@ class FamilyMemberForm(forms.ModelForm):
         model = Profile
         fields = ['title', 'first_name', 'last_name', 'names_inversed', 'birth_date']
         error_messages = {
-            'birth_date': { 'max_value': _("A family member cannot be future-born (even if planned)."), },
+            'birth_date': {'max_value': _("A family member cannot be future-born (even if planned).")},
         }
 
     def __init__(self, *args, **kwargs):
@@ -341,16 +302,3 @@ class FamilyMemberCreateForm(FamilyMemberForm):
         family_member = super(FamilyMemberCreateForm, self).save()
         self.place.family_members.add(family_member)
         return family_member
-
-
-class MassMailForm(forms.Form):
-    subject = forms.CharField(label=_("Subject"), initial="[Pasporta Servo] ")
-    body = forms.CharField(label=_("Body"), widget=forms.Textarea, initial="Kara {nomo},")
-    categories = forms.ChoiceField(label=_("Categories"), choices=(
-        ('test', _("test")),
-        ('old_system', _("old system")),
-        ('not_in_book', _("not in book")),
-        ('in_book', _("in book")),
-        ('just_user', _("just user")),
-    ))
-    test_email = forms.EmailField(label=_("Your email for test"), initial="baptiste@darthenay.fr")
